Выполнение задания 2.15
