+++ index.html
@@ -31,9 +31,9 @@
         </button>
       </div>
     </div>
-    <script type="module" src="api.js"></script>
-    <script type="module" src="render.js"></script>
+    <script type="module" src="./module_js/api.js"></script>
+    <script type="module" src="./module_js/render.js"></script>
   </body>
 
-  <script src="main.js"></script>
+  <script type="module" src="./module_js/main.js"></script>
 </html>
